meme_public_api/meme_app/config.py
import os

from dotenv import load_dotenv, find_dotenv

load_dotenv(find_dotenv())


class Config:
    DATABASE_URL = os.getenv('DATABASE_URL')
    MEDIA_API_URL = os.getenv('MEDIA_API_URL')

    POSTGRES_DB = os.getenv('POSTGRES_DB')
    POSTGRES_PORT = os.getenv('POSTGRES_PORT')
    POSTGRES_USER = os.getenv('POSTGRES_USER')
    POSTGRES_PASSWORD = os.getenv('POSTGRES_PASSWORD')


config = Config()
