meme_public_api/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==20.3.0
build==1.2.1
CacheControl==0.14.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
cryptography==42.0.8
distlib==0.3.8
dnspython==2.6.1
dulwich==0.21.7
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
fastjsonschema==2.19.1
filelock==3.14.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
installer==0.7.0
jaraco.classes==3.4.0
jellyfish==0.11.2
Jinja2==3.1.4
keyring==24.3.1
lib50==3.0.11
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
minio==7.2.7
more-itertools==10.2.0
msgpack==1.0.8
packaging==24.1
pexpect==4.9.0
pkginfo==1.11.1
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
psycopg2==2.9.9
ptyprocess==0.7.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
pyproject_hooks==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rapidfuzz==3.9.3
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
submit50==3.1.5
termcolor==1.1.0
tomlkit==0.12.5
trove-classifiers==2024.5.22
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
virtualenv==20.26.2
watchfiles==0.22.0
websockets==12.0
xattr==1.1.0
zipp==3.19.2
